refactor: コンポーネント内の処理をhooksに切り出し(#16)
